+++ changelog.md
@@ -20,7 +20,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Fixed
 
--
+- locations did show properly in events
 
 ## [3.4.0] - 2025-07-27